fast_api/requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-tavily
langchain-ollama
langchain-google-genai
tzdata
wikipedia
wikipedia-api
newspaper4k
newsapi-python
psycopg[binary]
ddgs
python-dotenv
lxml_html_clean
fastapi
uvicorn
pydantic-settings
openai
chromadb
sentence_transformers
transformers
torch
# Materials + Content Generation Dependencies
plotly==6.1.1
kaleido
moviepy
pillow
replicate
requests
numpy
langchain-anthropic
